Add files via upload
Water turbidity refers to the cloudiness or haziness of water caused by the presence of suspended particles, such as sediment, algae, and other organic or inorganic substances. The measurement of water turbidity is an important indicator of water quality, as high turbidity levels can indicate contamination or the presence of pollutants.

In my project, I have used the data collected about the aspects IoT (Internet of Things) technology to sense the water turbidity. IoT involves the use of internet-connected devices, such as sensors, to collect and transmit data from the physical world to digital systems. In this case, the sensors are used to measure the turbidity of water and send that data to a digital system, where it can be analyzed and used for forecasting.

The forecasting aspect of my project involves using historical data on water turbidity to make predictions about future levels. This involved a simple ARIMA model.